Limit the visibility of hd_key so it won't get misused outside of keyring
