+++ browser/ui/webui/brave_wallet/common_handler/BUILD.gn
@@ -11,6 +11,7 @@ source_set("common_handler") {
   deps = [
     "//brave/browser/brave_wallet",
     "//brave/components/brave_wallet/browser",
+    "//brave/components/brave_wallet/browser:hd_keyring",
     "//brave/components/brave_wallet/common:mojom",
     "//chrome/browser/profiles:profile",
     "//content/public/browser",
