+++ components/brave_wallet/browser/brave_wallet_utils.cc
@@ -32,10 +32,6 @@
 #include "third_party/boringssl/src/include/openssl/evp.h"
 #include "url/gurl.h"
 
-#if defined(OS_WIN)
-#include <windows.h>
-#endif
-
 namespace brave_wallet {
 
 namespace {
@@ -462,20 +458,6 @@ bool DecodeStringArray(const std::string& input,
   }
 
   return true;
-}
-
-void SecureZeroData(void* data, size_t size) {
-  if (data == nullptr || size == 0)
-    return;
-#if defined(OS_WIN)
-  SecureZeroMemory(data, size);
-#else
-  // 'volatile' is required. Otherwise optimizers can remove this function
-  // if cleaning local variables, which are not used after that.
-  volatile uint8_t* d = (volatile uint8_t*)data;
-  for (size_t i = 0; i < size; i++)
-    d[i] = 0;
-#endif
 }
 
 // Updates preferences for when the wallet is unlocked.
