+++ components/brave_wallet/browser/eip2930_transaction_unittest.cc
@@ -9,7 +9,7 @@
 #include "base/strings/string_util.h"
 #include "base/values.h"
 #include "brave/components/brave_wallet/browser/eip2930_transaction.h"
-#include "brave/components/brave_wallet/browser/hd_key.h"
+#include "brave/components/brave_wallet/browser/internal/hd_key.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
 namespace brave_wallet {
