+++ components/brave_wallet/browser/eth_transaction_unittest.cc
@@ -10,7 +10,7 @@
 #include "base/strings/string_util.h"
 #include "base/values.h"
 #include "brave/components/brave_wallet/browser/eth_transaction.h"
-#include "brave/components/brave_wallet/browser/hd_key.h"
+#include "brave/components/brave_wallet/browser/internal/hd_key.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
 namespace brave_wallet {
