+++ components/brave_wallet/browser/internal/hd_key_ed25519.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_wallet/browser/hd_key_ed25519.h"
+#include "brave/components/brave_wallet/browser/internal/hd_key_ed25519.h"
 
 #include <utility>
