+++ components/brave_wallet/browser/test/BUILD.gn
@@ -27,9 +27,9 @@ source_set("brave_wallet_unit_tests") {
     "//brave/components/brave_wallet/browser/ethereum_permission_utils_unittest.cc",
     "//brave/components/brave_wallet/browser/fil_requests_unittest.cc",
     "//brave/components/brave_wallet/browser/fil_response_parser_unittest.cc",
-    "//brave/components/brave_wallet/browser/hd_key_ed25519_unittest.cc",
-    "//brave/components/brave_wallet/browser/hd_key_unittest.cc",
     "//brave/components/brave_wallet/browser/hd_keyring_unittest.cc",
+    "//brave/components/brave_wallet/browser/internal/hd_key_ed25519_unittest.cc",
+    "//brave/components/brave_wallet/browser/internal/hd_key_unittest.cc",
     "//brave/components/brave_wallet/browser/json_rpc_response_parser_unittest.cc",
     "//brave/components/brave_wallet/browser/json_rpc_service_unittest.cc",
     "//brave/components/brave_wallet/browser/password_encryptor_unittest.cc",
@@ -45,6 +45,9 @@ source_set("brave_wallet_unit_tests") {
     "//base/test:test_support",
     "//brave/components/brave_wallet/browser",
     "//brave/components/brave_wallet/browser:ethereum_permission_utils",
+    "//brave/components/brave_wallet/browser:hd_keyring",
+    "//brave/components/brave_wallet/browser:transaction",
+    "//brave/components/brave_wallet/browser/internal:hd_key",
     "//brave/components/brave_wallet/common",
     "//brave/components/brave_wallet/common:mojom",
     "//brave/components/ipfs",
